+++ react/components/form/ArrayFieldTemplateItem.js
@@ -1,6 +1,7 @@
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 import { SortableElement, SortableHandle } from 'react-sortable-hoc'
+import { Transition } from 'react-spring'
 
 import TrashSimple from '../icons/TrashSimple'
 import DragHandle from '../icons/DragHandle'
@@ -24,6 +25,7 @@ class ArrayFieldTemplateItem extends Component {
     isOpen: PropTypes.bool,
     onOpen: PropTypes.func,
     onClose: PropTypes.func,
+    showDragHandle: PropTypes.bool,
   }
 
   handleLabelClick = e => {
@@ -36,6 +38,16 @@ class ArrayFieldTemplateItem extends Component {
     }
   }
 
+  renderChildren = styles => {
+    const { children } = this.props
+
+    return (
+      <div style={styles}>
+        {children}
+      </div>
+    )
+  }
+
   render() {
     const {
       children,
@@ -44,15 +56,16 @@ class ArrayFieldTemplateItem extends Component {
       hasRemove,
       onDropIndexClick,
       isOpen,
+      showDragHandle,
     } = this.props
 
     const title = children.props.formData.__editorItemTitle || schema.items.properties.__editorItemTitle.default
 
     return (
-      <div className="accordion-item bb b--light-silver">
+      <div className={`accordion-item bb b--light-silver ${showDragHandle ? '' : 'accordion-item--handle-hidden'}`}>
         <div className="accordion-label" onClick={this.handleLabelClick}>
           <div className="flex items-center">
-            <Handle />
+            {showDragHandle && <Handle />}
             <label className="f6 accordion-label-title">
               {title}
             </label>
@@ -63,13 +76,22 @@ class ArrayFieldTemplateItem extends Component {
                 className="accordion-icon-button accordion-icon-button--remove"
                 onClick={stopPropagation(onDropIndexClick(formIndex))}
               >
-                <TrashSimple size="15" />
+                <TrashSimple size={15} />
               </button>
             )}
           </div>
         </div>
-        <div className="accordion-content">
-          {isOpen && children}
+        <div
+          className={`accordion-content ${isOpen ? 'accordion-content--open' : ''}`}
+        >
+          <Transition
+            keys={isOpen ? ['children'] : []}
+            from={{ opacity: 0, height: 0 }}
+            enter={{ opacity: 1, height: 'auto' }}
+            leave={{ opacity: 0, height: 0 }}
+          >
+            {isOpen ? [this.renderChildren] : []}
+          </Transition>
         </div>
       </div>
     )